refactor: position 패키지 생성 및 관련 패키지 위치 변경
