+++ src/main/java/chess/domain/ChessMap.java
@@ -1,6 +1,7 @@
 package chess.domain;
 
 import chess.domain.piece.Piece;
+import chess.domain.position.Position;
 import java.util.List;
 
 public class ChessMap {
