+++ src/main/java/chess/domain/generator/Generator.java
@@ -1,6 +1,5 @@
 package chess.domain.generator;
 
-import chess.domain.Position;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
@@ -8,6 +7,7 @@ import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
+import chess.domain.position.Position;
 import java.util.ArrayList;
 import java.util.List;
