+++ src/main/java/chess/domain/player/Player.java
@@ -1,6 +1,6 @@
 package chess.domain.player;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 import chess.domain.generator.Generator;
 import chess.domain.piece.Piece;
 import java.util.ArrayList;
