+++ src/test/java/chess/domain/generator/BlackGeneratorTest.java
@@ -2,7 +2,7 @@ package chess.domain.generator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
